Some Refactoring.
