+++ src/test/libslic3r/models/overhangs.model
