+++ xs/src/libslic3r/SupportMaterial.hpp
@@ -1,21 +1,18 @@
 #ifndef slic3r_SupportMaterial_hpp_
 #define slic3r_SupportMaterial_hpp_
 
-#include "Polygon.hpp"
-#include "ExPolygon.hpp"
-#include "ClipperUtils.hpp"
-#include "PrintConfig.hpp"
-#include "Print.hpp"
-#include "EdgeGrid.hpp"
-#include "Geometry.hpp"
-
 #include <tbb/parallel_for.h>
 #include <tbb/atomic.h>
 #include <tbb/spin_mutex.h>
 #include <tbb/task_group.h>
 
-#define MIN_LAYER_HEIGHT 0.01
-#define MIN_LAYER_HEIGHT_DEFAULT 0.07
+#include "Polygon.hpp"
+#include "ExPolygon.hpp"
+#include "Print.hpp"
+#include "EdgeGrid.hpp"
+#include "Geometry.hpp"
+#include "ClipperUtils.hpp"
+#include "SupportParameters.hpp"
 
 #define SUPPORT_SURFACES_OFFSET_PARAMETERS ClipperLib::jtSquare, 0.
 
@@ -26,87 +23,6 @@ using namespace std;
 
 namespace Slic3r
 {
-
-// TODO @Samir55, Refactor later to be phased out if possible.
-class SupportParameters
-{
-public:
-    SupportParameters() { memset(this, 0, sizeof(SupportParameters)); }
-
-    static SupportParameters
-    create_from_config(
-        const PrintConfig &print_config,
-        const PrintObjectConfig &object_config,
-        coordf_t object_height,
-        const std::set<size_t> &object_extruders);
-
-    // Has any raft layers?
-    bool
-    has_raft() const { return raft_layers() > 0; }
-
-    size_t
-    raft_layers() const { return base_raft_layers + interface_raft_layers; }
-
-    // Is the 1st object layer height fixed, or could it be varied?
-    bool
-    first_object_layer_height_fixed() const { return !has_raft() || first_object_layer_bridging; }
-
-    // Height of the object to be printed. This value does not contain the raft height.
-    coordf_t
-    object_print_z_height() const { return object_print_z_max - object_print_z_min; }
-
-    // Number of raft layers.
-    size_t base_raft_layers;
-    // Number of interface layers including the contact layer.
-    size_t interface_raft_layers;
-
-    // Layer heights of the raft (base, interface and a contact layer).
-    coordf_t base_raft_layer_height;
-    coordf_t interface_raft_layer_height;
-    coordf_t contact_raft_layer_height;
-    bool contact_raft_layer_height_bridging;
-
-    // The regular layer height, applied for all but the first layer, if not overridden by layer ranges
-    // or by the variable layer thickness table.
-    coordf_t layer_height;
-    // Minimum / maximum layer height, to be used for the automatic adaptive layer height algorithm,
-    // or by an interactive layer height editor.
-    coordf_t min_layer_height;
-    coordf_t max_layer_height;
-    coordf_t max_suport_layer_height;
-
-    // First layer height of the print, this may be used for the first layer of the raft
-    // or for the first layer of the print.
-    coordf_t first_print_layer_height;
-
-    // Thickness of the first layer. This is either the first print layer thickness if printed without a raft,
-    // or a bridging flow thickness if printed over a non-soluble raft,
-    // or a normal layer height if printed over a soluble raft.
-    coordf_t first_object_layer_height;
-
-    // If the object is printed over a non-soluble raft, the first layer may be printed with a briding flow.
-    bool first_object_layer_bridging;
-
-    // Soluble interface? (PLA soluble in water, HIPS soluble in lemonen)
-    // otherwise the interface must be broken off.
-    bool soluble_interface;
-    // Gap when placing object over raft.
-    coordf_t gap_raft_object;
-    // Gap when placing support over object.
-    coordf_t gap_object_support;
-    // Gap when placing object over support.
-    coordf_t gap_support_object;
-
-    // Bottom and top of the printed object.
-    // If printed without a raft, object_print_z_min = 0 and object_print_z_max = object height.
-    // Otherwise object_print_z_min is equal to the raft height.
-    coordf_t raft_base_top_z;
-    coordf_t raft_interface_top_z;
-    coordf_t raft_contact_top_z;
-    // In case of a soluble interface, object_print_z_min == raft_contact_top_z, otherwise there is a gap between the raft and the 1st object layer.
-    coordf_t object_print_z_min;
-    coordf_t object_print_z_max;
-};
 
 // how much we extend support around the actual contact area
 constexpr coordf_t SUPPORT_MATERIAL_MARGIN = 1.5;
@@ -208,7 +124,7 @@ public:
         ///< Z used for printing, in unscaled coordinates.
 
         coordf_t bottom_z;
-        ///< Bottom Z of this layer. For soluble layers, bottom_z + height = print_z, TODO @Samir55 ASK.
+        ///< Bottom Z of this layer. For soluble layers, bottom_z + height = print_z, TODO @Samir55 ASK. [RESOLVED :)]
         ///< otherwise bottom_z + gap + height = print_z.
 
         coordf_t height;
@@ -226,7 +142,7 @@ public:
         ///< Use a bridging flow when printing this support layer.
 
         Polygons polygons;
-        // Polygons to be filled by the support pattern.
+        ///< Polygons to be filled by the support pattern.
 
         Polygons *contact_polygons;
         ///< Currently for the contact layers only.
@@ -246,25 +162,25 @@ public:
 
     /// Is raft enabled?
     bool
-    has_raft() const;
+    has_raft() const { return m_support_params.has_raft(); }
 
     /// Has any support?
     bool
-    has_support() const;
+    has_support() const { return m_object_config->support_material.value; }
 
     /// Build supports on plate only allowed
     bool
-    build_plate_only() const;
+    build_plate_only() const { return this->has_support() && m_object_config->support_material_buildplate_only.value; }
 
     ///
     /// \return
     bool
-    synchronize_layers() const;
+    synchronize_layers() const { return m_support_params.soluble_interface && m_object_config->support_material_synchronize_layers.value; }
 
     ///
     /// \return
     bool
-    has_contact_loops() const;
+    has_contact_loops() const { return m_object_config->support_material_interface_contact_loops.value; }
 
     /// Generate support material for the object.
     /// New support layers will be added to the object,
@@ -276,7 +192,7 @@ public:
 private:
     /// Generate top contact layers supporting overhangs. TODO @Samir55 ASK.
     /// For a soluble interface material synchronize the layer heights with the object, otherwise leave the layer height undefined.
-    /// If supports over bed surface only are requested, don't generate contact layers over an object.
+    /// If supports over bed surface only are requested, don't generate top contact layers over an object.
     /// \param object
     /// \param layer_storage
     /// \return
@@ -304,6 +220,7 @@ private:
     trim_top_contacts_by_bottom_contacts(const PrintObject &object, const MyLayersPtr &bottom_contacts,
                                          MyLayersPtr &top_contacts) const;
 
+    /// Called by
     /// Generate raft layers and the intermediate support layers between the bottom contact and top contact surfaces.
     /// \param object
     /// \param bottom_contacts
@@ -317,6 +234,7 @@ private:
         const MyLayersPtr &top_contacts,
         MyLayerStorage &layer_storage) const;
 
+    /// Called by
     /// Fill in the base layers with polygons.
     void
     generate_base_layers(
@@ -326,6 +244,7 @@ private:
         MyLayersPtr &intermediate_layers,
         const std::vector<Polygons> &layer_support_areas) const;
 
+    /// Called by
     /// Generate raft layers, also expand the 1st support layer
     /// in case there is no raft layer to improve support adhesion.
     /// \param top_contacts
@@ -340,6 +259,7 @@ private:
         const MyLayersPtr &base_layers,
         MyLayerStorage &layer_storage) const;
 
+    /// Called by
     /// Turn some of the base layers into interface layers.
     MyLayersPtr
     generate_interface_layers(
@@ -400,7 +320,7 @@ private:
     coordf_t m_gap_xy; ///<
 };
 
-/// Support grid pattern class. TODO @Samir Add contour based method also if applicable.
+/// Support grid pattern class.
 class SupportGridPattern
 {
 public:
